warning : size_t comparison against signed integer
git-svn-id: https://projets-systeme.lip6.fr/svn/research/libddd/libddd/trunk@1595 ab02c66d-5113-476d-a8b6-74e7035bd63a

+++ demo/morpion/hom/morpion_hom_vyann.cpp
@@ -180,7 +180,7 @@ class _Play:public StrongHom
     bool
         skip_variable(int vr) const
     {
-      return vr != cell && vr!=STATE_SYSTEM_CELL;
+      return vr != cell && size_t(vr)!=STATE_SYSTEM_CELL;
     }
 
     /**
@@ -203,7 +203,7 @@ class _Play:public StrongHom
     GHom
         phi(int vr, int vl) const
     {
-      if(vr!=STATE_SYSTEM_CELL)
+      if(size_t(vr)!=STATE_SYSTEM_CELL)
       {
         /* Configuration 1 : We can take a cell only if nobody get it*/
         if (vl == EMPTY)
@@ -321,7 +321,7 @@ class _NoteWinner:public StrongHom
     bool
         skip_variable(int vr) const
     {
-      return vr!=STATE_SYSTEM_CELL;
+      return size_t(vr) !=STATE_SYSTEM_CELL;
     }
 
     /**
